refactor(shared): rename asset_record_interface.py to asset_record_repository.py
インターフェース命名規則の整理。*_interface.py サフィックスを廃止し、
責務（repository）を表すサフィックスに変更。クラス名
IAssetRecordRepository は変更なし。
