+++ lambda/shared/src/shared/domain/asset_record_repository.py
